+++ src/App.js
@@ -22,7 +22,7 @@ import { getTheme } from 'store/theme'
 import { Helmet } from 'react-helmet'
 import { createBrowserHistory } from 'history'
 import ErrorBoundary from 'components/ErrorBoundary'
-import { getLastSyncTime, getRootUserId } from 'store/serverData'
+import { getLastSyncTime, getRootUserId } from 'store/data/selectors'
 import Accounts from 'scenes/Accounts'
 import Stats from 'scenes/Stats'
 import About from 'scenes/About'
