+++ src/components/RegularSyncHandler.js
@@ -2,7 +2,7 @@ import React from 'react'
 import { connect } from 'react-redux'
 import { getLoginState } from 'store/token'
 import { syncData } from 'logic/sync'
-import { getLastSyncTime } from 'store/serverData'
+import { getLastSyncTime } from 'store/data/selectors'
 import { getLastChangeTime } from 'store/localData'
 import { getPendingState } from 'store/isPending'
 import { loadLocalData } from 'logic/localData'
