+++ src/logic/exportJSON.ts
@@ -1,6 +1,6 @@
 import { formatDate } from 'helpers/format'
 import { AppThunk } from 'store'
-import { getDataToSave } from 'store/serverData'
+import { getDataToSave } from 'store/data/selectors'
 
 export const exportJSON: AppThunk = (_, getState) => {
   const state = getState()
