+++ src/scenes/Auth/index.js
@@ -6,7 +6,7 @@ import { logIn } from 'logic/authorization'
 import ZenApi from 'services/ZenApi'
 import { useTheme } from '@material-ui/styles'
 import Logo from '../../components/Logo'
-import { applyServerPatch } from 'store/dataSlice'
+import { applyServerPatch } from 'store/data'
 import { setToken } from 'store/token'
 import { saveDataLocally } from 'logic/localData'
 import { Link as RouterLink } from 'react-router-dom'
