+++ src/scenes/Budgets/containers/MonthInfo/WidgetIncome.js
@@ -3,7 +3,7 @@ import { useSelector } from 'react-redux'
 import { endOfMonth } from 'date-fns'
 import { getTotalsByMonth } from '../../selectors/getTotalsByMonth'
 import { getAmountsByTag } from '../../selectors/getAmountsByTag'
-import { getUserCurrencyCode } from 'store/serverData'
+import { getUserCurrencyCode } from 'store/data/selectors'
 import {
   Box,
   Typography,
