+++ src/scenes/Budgets/containers/TagTable/BudgetPopover.js
@@ -12,7 +12,7 @@ import CheckCircleIcon from '@material-ui/icons/CheckCircle'
 import { formatMoney } from 'helpers/format'
 import AmountInput from 'components/AmountInput'
 import { getAmountsById } from 'scenes/Budgets/selectors/getAmountsByTag'
-import { getUserCurrencyCode } from 'store/serverData'
+import { getUserCurrencyCode } from 'store/data/selectors'
 import { setOutcomeBudget } from 'scenes/Budgets/thunks'
 import { getGoals } from 'store/localData/hiddenData/goals'
 import { getGoalsProgress } from 'scenes/Budgets/selectors/goalsProgress'
