+++ src/scenes/Budgets/containers/TransferTable.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { connect } from 'react-redux'
 import { formatMoney } from 'helpers/format'
-import { getUserCurrencyCode } from 'store/serverData'
+import { getUserCurrencyCode } from 'store/data/selectors'
 import { makeStyles } from '@material-ui/core/styles'
 import {
   Box,
