+++ src/scenes/Budgets/selectors/getAmountsByTag.ts
@@ -1,7 +1,7 @@
 import { createSelector } from '@reduxjs/toolkit'
 import { getMainTag } from 'store/localData/transactions/helpers'
 import { getTagLinks } from 'store/localData/tags'
-import { convertCurrency } from 'store/serverData'
+import { convertCurrency } from 'store/data/selectors'
 import { getBudgetsByMonthAndTag } from 'store/localData/budgets'
 import { round } from 'helpers/currencyHelpers'
 import { getAccTagMap } from 'store/localData/hiddenData/accTagMap'
