+++ src/scenes/Budgets/thunks/index.js
@@ -1,4 +1,4 @@
-import { getRootUser } from 'store/serverData'
+import { getRootUser } from 'store/data/selectors'
 import { selectors } from 'store/localData/budgets/selectors'
 import { getPopulatedTag } from 'store/localData/tags'
 import { getAmountsByTag } from '../selectors/getAmountsByTag'
@@ -10,7 +10,7 @@ import { subMonths } from 'date-fns/esm'
 import { getGoalsProgress } from '../selectors/goalsProgress'
 import { GOAL_TYPES } from 'store/localData/hiddenData/constants'
 import { getGoals } from 'store/localData/hiddenData/goals'
-import { applyClientPatch } from 'store/dataSlice'
+import { applyClientPatch } from 'store/data'
 const { TARGET_BALANCE } = GOAL_TYPES
 
 export const moveFunds = (amount, source, destination, monthDate) => (
