+++ src/scenes/Review/cards/IncomeCard.tsx
@@ -3,7 +3,7 @@ import { Box, Typography, Chip } from '@material-ui/core'
 import { useSelector } from 'react-redux'
 import Rhythm from 'components/Rhythm'
 import { Amount } from 'components/Amount'
-import { getUserCurrencyCode } from 'store/serverData'
+import { getUserCurrencyCode } from 'store/data/selectors'
 import { getPopulatedTags } from 'store/localData/tags'
 import pluralize from 'helpers/pluralize'
 import { Card } from './Card'
