+++ src/scenes/Review/cards/OutcomeCard.tsx
@@ -4,7 +4,7 @@ import { useSelector } from 'react-redux'
 import { formatDate } from 'helpers/format'
 import Rhythm from 'components/Rhythm'
 import { Amount } from 'components/Amount'
-import { getInstruments } from 'store/serverData'
+import { getInstruments } from 'store/data/selectors'
 import { getPopulatedTags } from 'store/localData/tags'
 import { Card } from './Card'
 import { Transaction } from 'types'
