+++ src/scenes/Review/cards/PayeeByFrequencyCard.tsx
@@ -3,7 +3,7 @@ import { Typography } from '@material-ui/core'
 import { useSelector } from 'react-redux'
 import Rhythm from 'components/Rhythm'
 import { Amount } from 'components/Amount'
-import { getUserCurrencyCode } from 'store/serverData'
+import { getUserCurrencyCode } from 'store/data/selectors'
 import pluralize from 'helpers/pluralize'
 import { Card } from './Card'
 import { Stats } from '../selectors'
