+++ src/store/localData/accounts/index.ts
@@ -1,6 +1,6 @@
 import { createSelector } from '@reduxjs/toolkit'
 import { populate } from './populate'
-import { convertCurrency } from 'store/serverData'
+import { convertCurrency } from 'store/data/selectors'
 import { Account, AccountId, PopulatedAccount } from 'types'
 import { RootState } from 'store'
