+++ src/store/localData/hiddenData/thunks.ts
@@ -1,9 +1,9 @@
-import { getRootUser } from 'store/serverData'
+import { getRootUser } from 'store/data/selectors'
 import { makeDataAcc, makeDataReminder } from './helpers'
 import { getDataReminders, getDataAccountId } from './selectors'
 import { DataRemindeType } from './constants'
 import { AppDispatch, AppThunk, AppGetState } from 'store'
-import { applyClientPatch } from 'store/dataSlice'
+import { applyClientPatch } from 'store/data'
 
 export const setHiddenData = (type: DataRemindeType, data: any): AppThunk => (
   dispatch,
